fix: breaks activity tab into components
