+++ app/src/app/[lng]/[inventory]/data/[step]/[subsector]/page.tsx
@@ -1,68 +1,29 @@
 "use client";
 
-import MethodologyCard from "@/components/Cards/methodology-card";
-import SuggestedActivityCard from "@/components/Cards/suggested-activities-card";
-import AddActivityModalEnergyConsumption from "@/components/Modals/add-activity-energy-consumption-modal";
-import AddActivityModal from "@/components/Modals/add-activity-modal";
-import ChangeMethodology from "@/components/Modals/change-methodology";
-import DeleteActivityModal from "@/components/Modals/delete-activity-modal";
-import DeleteAllActivitiesModal from "@/components/Modals/delete-all-activities-modal";
-import ActivityTab from "@/components/Tabs/activity-tab";
-import HeadingText from "@/components/heading-text";
-import LoadingState from "@/components/loading-state";
+import ActivityTab from "@/components/Tabs/Activity/activity-tab";
 import { useTranslation } from "@/i18n/client";
 import { RootState } from "@/lib/store";
-import { ActivityDataAttributes } from "@/models/ActivityData";
 import { api } from "@/services/api";
-import { AddIcon, ArrowBackIcon, ChevronRightIcon } from "@chakra-ui/icons";
+import { ArrowBackIcon, ChevronRightIcon } from "@chakra-ui/icons";
 import {
-  Accordion,
-  AccordionButton,
-  AccordionIcon,
-  AccordionItem,
-  AccordionPanel,
   Box,
   Breadcrumb,
   BreadcrumbItem,
   BreadcrumbLink,
   Button,
   Icon,
-  IconButton,
   Link,
-  Popover,
-  PopoverArrow,
-  PopoverBody,
-  PopoverContent,
-  PopoverTrigger,
-  Radio,
-  RadioGroup,
-  Stack,
-  Switch,
   Tab,
   TabList,
-  TabPanel,
   TabPanels,
-  Table,
-  TableContainer,
   Tabs,
-  Tag,
-  TagLabel,
-  Tbody,
-  Td,
   Text,
-  Textarea,
-  Th,
-  Thead,
-  Tr,
   useDisclosure,
 } from "@chakra-ui/react";
 import { motion } from "framer-motion";
 import { useRouter } from "next/navigation";
 import { useEffect, useState } from "react";
-import { Trans } from "react-i18next";
-import { FaNetworkWired } from "react-icons/fa";
-import { FiTrash2 } from "react-icons/fi";
-import { MdMoreVert, MdOutlineHomeWork } from "react-icons/md";
+import { MdOutlineHomeWork } from "react-icons/md";
 import { useSelector } from "react-redux";
 
 function SubSectorPage({
@@ -72,106 +33,17 @@ function SubSectorPage({
 }) {
   const router = useRouter();
   const { t } = useTranslation(lng, "data");
-  const [isMethodologySelected, setIsMethodologySelected] = useState(false);
-  const [selectedMethodology, setSelectedMethodology] = useState("");
 
   const { data: inventory, isLoading: isInventoryLoading } =
     api.useGetInventoryQuery(inventoryId);
   const { data: userActivities, isLoading: areActivitiesLoading } =
     api.useGetActivityValuesQuery({ inventoryId, subSectorId: subsector });
 
-  console.log(userActivities);
-
-  const [isUnavailableChecked, setIsChecked] = useState<boolean>(false);
-
-  const [hasActivityData, setHasActivityData] = useState<boolean>(false);
-
-  const handleSwitch = (e: any) => {
-    setIsChecked(!isUnavailableChecked);
-  };
-
-  // TODO add to DB and get from there
-  const METHODOLOGIES = [
-    {
-      methodologyId: "1",
-      name: t("fuel-combustion-consumption"),
-      description: t("fuel-combustion-consuption-desciption"),
-      inputRequired: [t("total-fuel-consumed")],
-      disabled: false,
-    },
-    {
-      methodologyId: "2",
-      name: t("scaled-sample-data"),
-      description: t("scaled-sample-data-desc"),
-      inputRequired: [t("sample-fuel"), t("scaling-data")],
-      disabled: false,
-    },
-    {
-      methodologyId: "3",
-      name: t("modeled-data"),
-      description: t("modeled-data-desc"),
-      inputRequired: [t("modeled-fuel"), t("build-area")],
-      disabled: true,
-    },
-    {
-      methodologyId: "4",
-      name: t("direct-measure"),
-      description: t("direct-measure-desc"),
-      inputRequired: [t("emissions-data")],
-      disabled: false,
-    },
-  ];
-  const suggestedActivities = [
-    {
-      id: "1",
-      name: t("commercial-buildings"),
-    },
-    {
-      id: "2",
-      name: t("institutional-buildings"),
-    },
-    {
-      id: "3",
-      name: t("street-lighting"),
-    },
-  ];
-
-  const handleMethodologySelected = (methodologyId: string) => {
-    setSelectedMethodology(methodologyId);
-    setIsMethodologySelected(!isMethodologySelected);
-  };
-
-  const {
-    isOpen: isAddActivityModalOpen,
-    onOpen: onAddActivityModalOpen,
-    onClose: onAddActivityModalClose,
-  } = useDisclosure();
-
-  const {
-    isOpen: isAddActivityModalOpenEC,
-    onOpen: onAddActivityModalOpenEC,
-    onClose: onAddActivityModalCloseEC,
-  } = useDisclosure();
-
-  const {
-    isOpen: isChangeMethodologyModalOpen,
-    onOpen: onChangeMethodologyOpen,
-    onClose: onChangeMethodologyClose,
-  } = useDisclosure();
-
   const {
     isOpen: isDeleteActivitiesModalOpen,
     onOpen: onDeleteActivitiesModalOpen,
     onClose: onDeleteActivitiesModalClose,
   } = useDisclosure();
-
-  const {
-    isOpen: isDeleteActivityModalOpen,
-    onOpen: onDeleteActivityModalOpen,
-    onClose: onDeleteActivityModalClose,
-  } = useDisclosure();
-
-  const [selectedActivity, setSelectedActivity] = useState<string | null>(null);
 
   const [selectedTab, setSelectedTab] = useState(1); // sector ID (1/2/3)
 
@@ -191,32 +63,7 @@ function SubSectorPage({
     (state: RootState) => state.subsector.subsector,
   );
 
-  const changeMethodology = () => {
-    setSelectedMethodology("");
-    setIsMethodologySelected(false);
-    onChangeMethodologyClose();
-  };
-
-  const [deleteActivity, isDeleteActivityLoading] =
-    api.useDeleteActivityValueMutation();
-
-  const deleteAllActivities = () => {
-    if (areActivitiesLoading || userActivities.length === 0) {
-      onDeleteActivitiesModalClose();
-      return;
-    }
-
-    for (const activity of userActivities) {
-      deleteActivity({ inventoryId, activityValueId: activity.id });
-    }
-
-    onDeleteActivitiesModalClose();
-  };
-
   // calculate total consumption and emissions
-  const totalConsumption = 123.3;
-  const totalConsumptionUnit = "M gallons";
-  const totalEmissions = 11.465; // MtCO2e
 
   const [scrollPosition, setScrollPosition] = useState<number>(0);
 
@@ -403,33 +250,6 @@ function SubSectorPage({
             </TabPanels>
           </Tabs>
         </Box>
-        <AddActivityModal
-          t={t}
-          userInfo={null}
-          isOpen={isAddActivityModalOpen}
-          onClose={onAddActivityModalClose}
-          hasActivityData={hasActivityData}
-          setHasActivityData={setHasActivityData}
-        />
-        <AddActivityModalEnergyConsumption
-          t={t}
-          userInfo={null}
-          isOpen={isAddActivityModalOpenEC}
-          onClose={onAddActivityModalCloseEC}
-          hasActivityData={hasActivityData}
-          setHasActivityData={setHasActivityData}
-        />
-        <ChangeMethodology
-          t={t}
-          onClose={onChangeMethodologyClose}
-          isOpen={isChangeMethodologyModalOpen}
-          onChangeClicked={changeMethodology}
-        />
-        <DeleteAllActivitiesModal
-          t={t}
-          isOpen={isDeleteActivitiesModalOpen}
-          onClose={onDeleteActivitiesModalClose}
-        />
       </div>
     </>
   );
